Abstract away big chunks of the DockerPipeline*Scripts.
