+++ pipeline-model-definition/src/main/resources/org/jenkinsci/plugins/pipeline/modeldefinition/agent/impl/DockerPipelineScript.groovy
@@ -25,52 +25,18 @@
 
 package org.jenkinsci.plugins.pipeline.modeldefinition.agent.impl
 
-import hudson.FilePath
 import hudson.model.Result
 import org.jenkinsci.plugins.pipeline.modeldefinition.SyntheticStageNames
 import org.jenkinsci.plugins.pipeline.modeldefinition.Utils
-import org.jenkinsci.plugins.pipeline.modeldefinition.agent.DeclarativeAgentScript
-import org.jenkinsci.plugins.pipeline.modeldefinition.steps.DeclarativePropsStep
 import org.jenkinsci.plugins.workflow.cps.CpsScript
 
-public class DockerPipelineScript extends DeclarativeAgentScript<DockerPipeline> {
+public class DockerPipelineScript extends AbstractDockerPipelineScript<DockerPipeline> {
 
     public DockerPipelineScript(CpsScript s, DockerPipeline a) {
         super(s, a)
     }
 
     @Override
-    public Closure run(Closure body) {
-        if (describable.reuseNode && script.getContext(FilePath.class) != null) {
-            return {
-                configureRegistry(body).call()
-            }
-        } else {
-            String targetLabel = script.declarativeProps(property: DeclarativePropsStep.Property.LABEL,
-                override: describable.label)
-            LabelScript labelScript = (LabelScript) Label.DescriptorImpl.instanceForName("label", [label: targetLabel]).getScript(script)
-            return labelScript.run {
-                configureRegistry(body).call()
-            }
-        }
-    }
-
-    private Closure configureRegistry(Closure body) {
-        return {
-            String registryUrl = script.declarativeProps(property: DeclarativePropsStep.Property.REGISTRY_URL,
-                override: describable.registryUrl)
-            String registryCreds = script.declarativeProps(property: DeclarativePropsStep.Property.REGISTRY_CREDENTIALS,
-                override: describable.registryCredentialsId)
-            if (registryUrl != null) {
-                script.getProperty("docker").withRegistry(registryUrl, registryCreds) {
-                    runImage(body).call()
-                }
-            } else {
-                runImage(body).call()
-            }
-        }
-    }
-
     public Closure runImage(Closure body) {
         return {
             if (!Utils.withinAStage()) {
